+++ functions.php
@@ -183,17 +183,3 @@ function excerpt($limit) {
   $excerpt = preg_replace('`[[^]]*]`','',$excerpt);
   return $excerpt;
 }
-
-function content($limit) {
-  $content = explode(' ', get_the_content(), $limit);
-  if (count($content)>=$limit) {
-    array_pop($content);
-    $content = implode(" ",$content).'...<a class="moretag" href="'. get_permalink($post->ID) . '"> continue reading &raquo;</a>';
-  } else {
-    $content = implode(" ",$content);
-  }
-  $content = preg_replace('/[.+]/','', $content);
-  $content = apply_filters('the_content', $content);
-  $content = str_replace(']]>', ']]&gt;', $content);
-  return $content;
-}