+++ index.php
@@ -19,29 +19,24 @@ get_header(); ?>
   <div class="container">
     <div class="row">
       <div class="col-md-12 section-title">
-        <h3>BLOG</h3>
+        <h3>Blog</h3>
       </div>
     </div>
 
 <main class="row" id="primary" role="main">
+  <?php
+
+     // the query
+     $the_query = new WP_Query( array(
+        'post_type' => 'post',
+        'posts_per_page' => 9,
+     ));
+  ?>
 
 		<?php if ( have_posts() ) :
-					if ( is_home() && ! is_front_page() ) : ?>
-				<header>
-					<h1 class="page-title screen-reader-text"><?php single_post_title(); ?></h1>
-				</header>
-				<?php endif;
-
-			/* Start the Loop */
-			while ( have_posts() ) : the_post();
-				/*
-				 * Include the Post-Format-specific template for the content.
-				 * If you want to override this in a child theme, then include a file
-				 * called content-___.php (where ___ is the Post Format name) and that will be used instead.
-				 */
+          while ( have_posts() ) : the_post();
 				get_template_part( 'template-parts/content', get_post_format() );
 			endwhile;
-			the_posts_navigation();
 		  else :
       get_template_part( 'template-parts/content', 'none' );
 		endif; ?>
@@ -49,25 +44,8 @@ get_header(); ?>
 
 </main><!--primary-->
 
-<nav aria-label="pagination navigation">
-  <ul class="pagination pagination-lg justify-content-center">
-    <li class="page-item disabled">
-      <a class="page-link" href="" aria-label="Previous">
-        <span aria-hidden="true">&laquo;</span>
-        <span class="sr-only">Previous</span>
-      </a>
-    </li>
-    <li class="page-item"><a class="page-link" href="#">1</a></li>
-    <li class="page-item"><a class="page-link" href="#">2</a></li>
-    <li class="page-item"><a class="page-link" href="#">3</a></li>
-    <li class="page-item">
-      <a class="page-link" href="" aria-label="Next">
-        <span aria-hidden="true">&raquo;</span>
-        <span class="sr-only">Next</span>
-      </a>
-    </li>
-  </ul>
-</nav>
+<?php get_template_part('template-parts/pagination', 'none' ); ?>
+
 
 </div><!--containter-->
 </section>