+++ template-parts/pagination.php
@@ -1,0 +1,13 @@
+<div class="row mt-4">
+  <div class="col-md-12">
+    <nav class="navigation" aria-label="pagination navigation">
+      <ul class="pagination pagination-lg justify-content-center">
+        <li class="page-item previous"><?php previous_post_link('&laquo;'); ?></li>
+        <li class="page-item"><a class="page-link" href="#">1</a></li>
+        <li class="page-item"><a class="page-link" href="#">2</a></li>
+        <li class="page-item"><a class="page-link" href="#">3</a></li>
+        <li class="page-item next"><?php next_post_link('&raquo;'); ?></li>
+      </ul>
+    </nav>
+  </div>
+</div>